Updates tapir dependency to 0.19.0-M16
Tapir has introduced the idea of an authentication stage, which can
extract auth and user contenxt. This caused a change to the signature
of a number of types, including `ServerEndpoint`.

This commit tries to provide the least amount of API change, that
allows "public" endpoints to be traced. Adding authenticated endpoints
requires more decisions, e.g. is there a need for
`ResourceKleisli[F, A, Either[E, Ctx]]`? How should the authentication
stage be traced, if at all?